% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_AML}
\alias{data_AML}
\title{Acute Myeloid Leukemia (AML) RNA-Seq Data}
\format{
A list with two components:
\describe{
  \item{x}{A numeric matrix of 22,834 gene expression levels (ultra-high dimensional).}
  \item{y}{A data frame with 444 observations on 2 variables:
    \itemize{
      \item \code{eln_risk}: 1 for high-risk, 0 for non-high-risk (ELN2017 criteria).
      \item \code{de_novo}: 1 for de novo AML, 0 for secondary or other types.
    }
  }
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=aml_ohsu_2022}
}
\usage{
data(data_AML)
}
\description{
Cleaned RNA-Seq expression profiles of 444 patients with Acute Myeloid Leukemia,
providing two clinical response variables for testing.
}
\references{
Bottomly D, Long N, Schultz A R, et al. Integrative analysis of drug response and clinical outcome in acute myeloid leukemia[J]. Cancer cell, 2022, 40(8): 850-864. e9.
}
\keyword{datasets}
