% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gof_test_fun.R
\name{Gof_Pcvm_test}
\alias{Gof_Pcvm_test}
\title{"Asymptotic Distribution-Free Tests for Ultra-high Dimensional Parametric Regressions via Projected Empirical Processes and p-value Combination"}
\usage{
Gof_Pcvm_test(x, y, family = c("gaussian", "binomial"))
}
\arguments{
\item{X}{Input matrix with \code{n} rows, each a \code{p}-dimensional observation vector.}

\item{Y}{Response vector.}

\item{fam}{Family type for GLM model fitting. Must be "gaussian" or "binomial".}
}
\value{
A list containing various aggregated p-values, where \code{pval_cauchy_hybrid} is the
primary result representing the combination of both Martingale and PLS tests.
}
\description{
This function constructs a Cramer-von Mises type test based on a martingale-transformed,
projected residual-marked empirical process. Furthermore, it proposes a novel hybrid test that aggregates empirical process-based tests and local smoothing tests using Cauchy combination.
}
\examples{
# Simulation example # H11
set.seed(123)
n <- 300
p <- 50
pho <- 0.4
mu <- rep(0, p)
v <- pho^(0:(p - 1))
sigma <- toeplitz(v)
x <- mvrnorm(n, mu, sigma)
beta0 <- c(rep(1, 5), rep(0, p - 5))

a <- 0
y <- x \%*\% beta0 + a * 0.1 * (x \%*\% beta0)^2 + rnorm(n)
output12 <- Gof_Pcvm_test(x, y, family = "gaussian")
a <- 1
y <- x \%*\% beta0 + a * 0.1 * (x \%*\% beta0)^2 + rnorm(n)
output14 <- Gof_Pcvm_test(x, y, family = "gaussian")

# Real data analysis
dim(data_crime)
y <- data_crime[, 100]
x <- data_crime[, -100]
output16 <- Gof_Pcvm_test(x, y, family = "gaussian")

# Simulation example  # H21
set.seed(123)
n <- 300
p <- 50
pho <- 0.4
mu <- rep(0, p)
v <- pho^(0:(p - 1))
sigma <- toeplitz(v)
x <- mvrnorm(n, mu, sigma)
beta0 <- c(rep(1, 5), rep(0, p - 5))

a <- 0
z <- x \%*\% beta0 + a * 0.2 * (x \%*\% beta0)^2
pr <- 1 / (1 + exp(-z))
y <- matrix(rbinom(n, 1, pr), ncol = 1)
output22 <- Gof_Pcvm_test(x, y, family = "binomial")
a <- 1
z <- x \%*\% beta0 + a * 0.2 * (x \%*\% beta0)^2
pr <- 1 / (1 + exp(-z))
y <- matrix(rbinom(n, 1, pr), ncol = 1)
output24 <- Gof_Pcvm_test(x, y, family = "binomial")

# Real data analysis
dim(data_AML)
y <- data_AML$ELN_binary
x <- data_AML[, 1:(ncol(data_AML) - 2)]
x <- x[, -which(is.na(apply(data_AML, 2, sum)))]
output26 <- Gof_Pcvm_test(x, y, family = "binomial")

}
